% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI_MCMC.R
\name{LOG_LIKE_SSI}
\alias{LOG_LIKE_SSI}
\title{Log likelihood of the Super-Spreading Individuals (SSI) epidemic model}
\usage{
LOG_LIKE_SSI(sim_data, aX, bX, cX)
}
\arguments{
\item{aX}{SSI model parameter \code{"a"}}

\item{bX}{SSI model parameter \code{"b"}}

\item{cX}{SSI model parameter \code{"c"}}

\item{data}{data from the epidemic, namely daily infection counts}
}
\value{
Log likelihood of the SSI model
}
\description{
Returns the Log likelihood of the super-spreading individuals (SSI)epidemic model for given epidemic \code{"data"} and set of model parameters \code{"a, b"} and \code{"c"} and
}
\examples{

log_like = LOG_LIKE_SSI(epidemic_data, 0.8, 0.02, 20)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
