% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMULATE_EPIDEMICS.R
\name{SIMULATE_BASELINE_EPIDEMIC}
\alias{SIMULATE_BASELINE_EPIDEMIC}
\title{Simulate an epidemic outbreak}
\usage{
SIMULATE_BASELINE_EPIDEMIC(R0, num_days = 50, shape_gamma = 6, scale_gamma = 1)
}
\arguments{
\item{R0}{Model parameter \code{"R0"}. The daily infection count from regular spreading is assumed to follow a poisson distribution with rate \code{R0}*\code{\lambda_t}}

\item{num_days}{Number of days of the epidemic}

\item{shape_gamma}{Shape of the gamma distribution representing the time-varying infectivity curve of each infected individual}

\item{scale_gamma}{Scale of the gamma distribution representing the time-varying infectivity curve of each infected individual}
}
\value{
Total infections; Total daily infection counts of length \code{'num_days'}
}
\description{
Returns daily infection counts from a simulated epidemic outbreak
}
\examples{

tot_daily_infections = SIMULATE_BASELINE_EPIDEMIC(num_days = 50, shape_gamma = 6, scale_gamma = 1, R0 = 1.2)S

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
